Move from typeclass methods to GADTs
This is a complete rewrite, barring the type level stuff. Form has been
removed in favour of a GADT which can hold some contexts according to
the 'Args' type family. Validation no longer requires orphan instances:
the validation functions now live inside the 'Form' GADT, when tagged
with 'ValFunc'. The 'EmptySetErrors' contexts are ugly but I think
necessary, this is something to be worked out.

ToJSON/FromJSON instances for 'Form Raw' are working (but untested,
indeed all the tests need rewriting). For the raw forms we try to encode
in the most "sensible" manner for a forms raw data: just use a JSON
object mapping field names to values.  Still need to provide To/FromJSON
for 'Form Err'.